+++ README.md
@@ -4,6 +4,12 @@
 - Installare git, cursor e le estensioni necessarie (lego spike prime, python, codeStackr theme)
 - Clonare questa repo col comando 'git clone'
 - Modificare i file e aggiornare la repo coi comandi 'git add .' 'git commit -m "specificare modifica effettuata" ' 'git push origin main' e accedere con l'account di Ciro su github
+## Missioni:
+### Da Fare/concludere:
+- 7
+- 8
+### Possibilmente da sistemare: 
+- 6
 ## To-do list:
 - [X] Calibrare il pid con il metodo Zieger-Nichols (scartato)
 - [x] Controllare i parametri e i valori nella funzione 'calcoloPID()'
